Incluido Get, Put, Delete na FornecedorController, Services. Criado Pastas ViewsModels

+++ SistemaVendas.Api/Services/ClienteService.cs
@@ -14,23 +14,35 @@ public class ClienteService(ApiDbContext context)
         return cliente;
     }
     
-    public async Task<Cliente> Update(Cliente cliente)
+    public async Task<Cliente> Update(int id, Cliente model)
     {
+        var cliente = await context.Clientes.FirstOrDefaultAsync(x=>x.Id == id);
+        if(cliente == null)
+            throw new InvalidOperationException("Cliente não encontrado");
+        cliente.AtualizarCliente(model.Nome, model.Telefone);
         context.Clientes.Update(cliente);
         await context.SaveChangesAsync();
 
         return cliente;
     }
     
-    public async Task Delete(Cliente cliente)
+    public async Task Delete(int id)
     {
+        var cliente = await context.Clientes.FirstOrDefaultAsync(x=>x.Id == id);
+        if(cliente == null)
+            throw new InvalidOperationException("Cliente não encontrado");
+        
         context.Clientes.Remove(cliente);
         await context.SaveChangesAsync();
     }
     
-    public async Task<Cliente?> GetById(int id)
+    public async Task<Cliente> GetById(int id)
     {
-        return await context.Clientes.FirstOrDefaultAsync(x=>x.Id == id);
+        var cliente = await context.Clientes.FirstOrDefaultAsync(x=>x.Id == id);
+        if(cliente == null)
+            throw new InvalidOperationException("Cliente não encontrado");
+        
+        return cliente;
     }
     
     public async Task<IEnumerable<Cliente>> GetAll()
